Prepare functions for reuse in other test.
