+++ tests/search/struct_and_map_types/struct_and_map_types.rb
@@ -1,7 +1,10 @@
 # Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 require 'indexed_streaming_search_test'
+require 'search/struct_and_map_types/struct_and_map_base'
 
 class StructAndMapTypesTest < IndexedStreamingSearchTest
+
+  include StructAndMapBase
 
   def setup
     set_owner("geirst")
@@ -135,35 +138,6 @@ class StructAndMapTypesTest < IndexedStreamingSearchTest
     assert_same_element_single_summary("elem_map_2", "key contains '@bar'", "default", "elem_map_2", {"@bar" => elem("bar", 20)})
     assert_same_element_single_summary("str_int_map", "key contains '@bar'", "default", "str_int_map", {"@foo" => 10, "@bar" => 20, "@baz" => 30})
     assert_same_element_single_summary("str_int_map", "key contains '@bar'", "filtered", "str_int_map_filtered", {"@bar" => 20})
-  end
-
-  def assert_same_element_summary(field, same_element, summary, summary_field, exp_summary_field)
-    form = [["yql", "select * from sources * where #{field} contains sameElement(#{same_element});"],
-            ["summary", summary ],
-            ["format", "json" ],
-            ["streaming.selection", "true"],
-            ["hits", "10"]]
-    assert_summary_field(form, summary_field, exp_summary_field)
-  end
-
-  def assert_same_element_single_summary(field, same_element, summary, summary_field, exp_summary_field)
-    form = [["yql", "select * from sources * where #{field}.#{same_element};"],
-            ["summary", summary ],
-            ["format", "json" ],
-            ["streaming.selection", "true"],
-            ["hits", "10"]]
-    assert_summary_field(form, summary_field, exp_summary_field)
-  end
-
-  def assert_summary_field(form, summary_field, exp_summary_field)
-    encoded_form = URI.encode_www_form(form)
-    puts "form is #{encoded_form}"
-    result = search("#{encoded_form}")
-    puts "result is #{result.xmldata}"
-    assert_equal(1, result.hitcount)
-    hit = result.hit[0]
-    act_summary_field = hit.field[summary_field]
-    assert_equal(exp_summary_field, act_summary_field)
   end
 
   def assert_same_element(field, same_element, exp_hitcount, extra_params = "")
